Move methods from crypto to otherCurrency
Move methods that are not related to Crypto class
