+++ src/crypto.js
@@ -90,44 +90,6 @@ class Crypto {
     return result.Data;
   }
 
-  async nbuExchange() {
-    const data = await safeSpawn('python', './src/parser.py');
-    console.table(data);
-  }
-
-  async monoExchange() {
-    const data = await safeGet('https://api.monobank.ua/bank/currency');
-    if (data.errorDescription) {
-      console.log(data.errorDescription);
-      return;
-    }
-    for (const curr of data) {
-      curr.currencyCodeA = codesList[curr.currencyCodeA];
-      curr.currencyCodeB = codesList[curr.currencyCodeB];
-      const rawDate = new Date(curr.date * 1000);
-      curr.date = `${rawDate.getDate()}` +
-        `.${rawDate.getMonth() + 1}` +
-        `.${rawDate.getFullYear()}`;
-    }
-    console.table(data);
-  }
-
-  async privatExchange() {
-    const cash = await safeGet('https://api.privatbank.ua/p24api/pubinfo?json&exchange&coursid=5');
-    const nonCash = await safeGet('https://api.privatbank.ua/p24api/pubinfo?exchange&json&coursid=11');
-    const rateTypes = [cash, nonCash];
-
-    while (true) {
-      const first = 'Do you want to get cash rate (1) or non-cash rate (2)?\n';
-      const userChoice = (await promised.question(first) - 1);
-      if (userChoice <= 1) console.table(rateTypes[userChoice]);
-
-      const second = 'Would you like to get another rate? (y/n)\n';
-      const option = await promised.question(second);
-      if (option !== 'y') break;
-    }
-  }
-
   async cryptoNews() {
     const info = await safeGet(`${this.defaultUrl} + 'v2/news/?lang=EN`);
     const data = info.Data;
