+++ src/otherCurrency.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { question } = require('./promised.js');
+const promised = require('./promised.js');
 const { Wallet } = require('./wallet.js');
 const { writeFile } = require('./crypto.js');
 
@@ -34,6 +34,45 @@ const btcAdrBalance = async () =>  {
   await writeFile(res);
   return;
 };
+
+const nbuExchange = async () =>  {
+  const data = await safeSpawn('python', './src/parser.py');
+  console.table(data);
+};
+
+const monoExchange = async () => {
+  const data = await safeGet('https://api.monobank.ua/bank/currency');
+  if (data.errorDescription) {
+    console.log(data.errorDescription);
+    return;
+  }
+  for (const curr of data) {
+    curr.currencyCodeA = codesList[curr.currencyCodeA];
+    curr.currencyCodeB = codesList[curr.currencyCodeB];
+    const rawDate = new Date(curr.date * 1000);
+    curr.date = `${rawDate.getDate()}` +
+        `.${rawDate.getMonth() + 1}` +
+        `.${rawDate.getFullYear()}`;
+  }
+  console.table(data);
+};
+
+const privatExchange = async () => {
+  const cash = await safeGet('https://api.privatbank.ua/p24api/pubinfo?json&exchange&coursid=5');
+  const nonCash = await safeGet('https://api.privatbank.ua/p24api/pubinfo?exchange&json&coursid=11');
+  const rateTypes = [cash, nonCash];
+
+  while (true) {
+    const first = 'Do you want to get cash rate (1) or non-cash rate (2)?\n';
+    const userChoice = (await promised.question(first) - 1);
+    if (userChoice <= 1) console.table(rateTypes[userChoice]);
+
+    const second = 'Would you like to get another rate? (y/n)\n';
+    const option = await promised.question(second);
+    if (option !== 'y') break;
+  }
+};
+
 
 module.exports = {
   genWalletFeature,
